plotter, autorange
plotter, autorange

autorange mode for plotter
Automatic estimation of noisefloor in spectrum, min for waterfall and panadapter is adjusted automatically, max uses the old delta (max-min).

- double click y-axis to enable/disable
- y-axis turns green
- autorange mode is rememberd in settings

Noise Floor detection a la Simon Brown, quote:
"A few weeks previously a reasonable logic was implemented for measuring the noise floor.
Purists will not be happy - they rarely are, but it works for me.
Take the output from the SDR radio, ignore 15% of the bandwidth at the high and low end of the output to avoid the ant-alias filtering,
and we're left with a healthy 70% of the signal.
Now sort the FFT bins by value, take the mean of the lowest 10% and that's the noise floor."

Percent values are slightly different, a moving average is added over lowest bins and over time.

Update plotter.cpp